If you get a coin you will get 1 sec extra !

+++ Assets/Scripts/TimeManager.cs
@@ -1,11 +1,9 @@
 ﻿using System;
 
-using System;
-
 public class TimeManager : Singleton<TimeManager> {
 	protected TimeManager () {} 
 
-	public int time = 100;
+	public float time = 100f;
 }
